requirements.txt
pandas
numpy 
python-dotenv
pymysql
mysql-connector-python
scikit-learn
matplotlib 
seaborn
XGBOOST
catboost
mlflow==2.2.2